+++ test/minimal/ngssc.sh
@@ -1,3 +1,2 @@
 #!/bin/sh
 ngssc insert /usr/share/nginx/html -r
-nginx -g 'daemon off;'